libusb-sys: build native libusb on FreeBSD
FreeBSD maintains a libusb that is compatible with the libusb used
everywhere else. Upstream for libusb-sys uses pkgconf for this, which
'just works' and grabs -lusb from the base system.

wishbone-tool vendored libusb, so add the build glue for FreeBSD libusb to
the build. With this, wishbone-tool can successfully build.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>