wishbone-tool: config: check for uart keys if csr-csv passed
Previously, we would only validate the `uart_xover` keys were present if
the user specified a GDB server.  This could result in a situation where
the user provided a csr-csv without a uart, but still wanted to run a
terminal.

In this case, `wishbone-tool` would fall back on its default, hardcoded
values of 0xe0001800, which is incorrect.

With this patch, specifying `-s terminal --csr-csv [file.csv]` when the
csv file has no `uart_xover` registers will now throw an error.

Furthermore, with this patch, the `-s gdb` option required a uart, which
would prevent a user from opening a gdb server on a device without a
uart.

This fixes #26.

Signed-off-by: Sean Cross <sean@xobs.io>