config: apply register offsets to ALL registers
Rename `--csr-csv-offset` to `--register-offset` and ensure that this
offset is applied to ALL addresses.  This can be used to work with a
bridge without needing to manually adjust offsets.

This necessitated being able to store `None` inside the register map so
we can know if a csr lookup was valid-but-out-of-range.  Now if a user
gives a text-based lookup and it exists in csr.csv but is not mappable,
an error is generated rather than using a [potentially incorrect]
default.

Signed-off-by: Sean Cross <sean@xobs.io>